Feat: Created AddItemButton component. Added AddItemButton to '/Dashboard'.
